move availability API classes in a separate package
